fix: moved shared data to parent component and managed via props
